# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: codexws HTTP API
  version: 3.0.0
  description: 'Current same-origin API for the one shared codexws backend process.

    This contract documents deployed paths; it is not the proposed /api/v1 design.

    '
servers:
- url: /
...
paths:
...
  /api/workspace-files:
    get:
      tags:
      - files
      operationId: listWorkspaceFiles
      parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - name: dir
        in: query
        schema:
          type: string
          default: ''
      responses:
        '200':
          description: Non-hidden entries in a workspace directory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListing'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/workspace-files/read:
    get:
      tags:
      - files
      operationId: readWorkspaceFile
      parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/FilePath'
      responses:
        '200':
          description: UTF-8 text file content and metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileContent'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/workspace-files/write:
    post:
      tags:
      - files
      operationId: writeWorkspaceFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/WorkspaceFileSelector'
              - type: object
                required:
                - content
                properties:
                  content:
                    type: string
      responses:
        '200':
          description: File was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMutationResult'
        '400':
          $ref: '#/components/responses/Error'
  /api/workspace-files/create:
    post:
      tags:
      - files
      operationId: createWorkspaceFileEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/WorkspaceFileSelector'
              - type: object
                properties:
                  is_dir:
                    type: boolean
                    default: false
      responses:
        '200':
          description: Empty file or directory was created.
          content:
            application/json:
              schema:
                type: object
                required:
                - ok
                properties:
                  ok:
                    type: boolean
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/workspace-files/upload:
    post:
      tags:
      - files
      operationId: uploadWorkspaceFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - workspace_id
              - file
              properties:
                workspace_id:
                  type: string
                dir:
                  type: string
                  default: ''
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Uploaded file metadata.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/FileMutationResult'
                - type: object
                  properties:
                    path:
                      type: string
                    name:
                      type: string
        '400':
          $ref: '#/components/responses/Error'
  /api/workspace-files/download:
    get:
      tags:
      - files
      operationId: downloadWorkspaceFile
      parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/FilePath'
      responses:
        '200':
          description: File attachment.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/workspace-files/open-folder:
    post:
      tags:
      - files
      operationId: openWorkspaceFolder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - workspace_id
              properties:
                workspace_id:
                  type: string
                dir:
                  type: string
                  default: ''
      responses:
        '200':
          description: Native file manager launch was requested.
          content:
            application/json:
              schema:
                type: object
                required:
                - ok
                - path
                properties:
                  ok:
                    type: boolean
                  path:
                    type: string
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    WorkspaceId:
      name: workspace_id
      in: query
      required: true
      schema:
        type: string
        minLength: 1
...
    FilePath:
      name: path
      in: query
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
...
    FileEntry:
      type: object
      required:
      - name
      - path
      - is_dir
      - size
      - modified
      properties:
        name:
          type: string
        path:
          type: string
        is_dir:
          type: boolean
        size:
          type: integer
          minimum: 0
        modified:
          type: string
    FileListing:
      type: object
      required:
      - workspace_id
      - files
      - current_dir
      properties:
        workspace_id:
          type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileEntry'
        current_dir:
          type: string
        parent_dir:
          type:
          - string
          - 'null'
    FileContent:
      type: object
      required:
      - content
      - size
      - modified
      properties:
        content:
          type: string
        size:
          type: integer
          minimum: 0
        modified:
          type: string
    WorkspaceFileSelector:
      type: object
      required:
      - workspace_id
      - path
      properties:
        workspace_id:
          type: string
          minLength: 1
        path:
          type: string
          minLength: 1
    FileMutationResult:
      type: object
      required:
      - ok
      properties:
        ok:
          type: boolean
        size:
          type: integer
          minimum: 0
        modified:
          type: string